[adjustments][pat] add routes in constants and changed components and pages structure
