+++ src/App.js
@@ -1,20 +1,15 @@
-import { lazy, Suspense, React } from 'react';
+import { Suspense, React } from 'react';
 import { Route, BrowserRouter as Router, Switch, Redirect } from 'react-router-dom';
 
-// Pages/Routes
-const LoginSignup = lazy(() => import('pages/LoginSignup'));
+import * as routes from 'constants/routes';
 
 function App() {
   return (
     <Router>
       <Suspense fallback={<div>Loading...</div>}>
         <Switch>
-          <Route path="/login">
-            <LoginSignup />
-          </Route>
-          <Route path="/signup">
-            <LoginSignup />
-          </Route>
+          <Route path="/login" component={routes.Login} />
+          <Route path="/signup" component={routes.Signup} />
           <Redirect to="/login" />
         </Switch>
       </Suspense>
