+++ src/constants/routes.js
@@ -1,0 +1,4 @@
+import { lazy } from 'react';
+
+export const Login = lazy(() => import('pages/Login'));
+export const Signup = lazy(() => import('pages/Signup'));
