+++ src/pages/Login/LoginPage.scss
@@ -1,4 +1,16 @@
-@import '../../../utils/variables.scss';
+@import '../../utils/variables.scss';
+
+.login {
+  background-color: $primary-background;
+  height: 100vh;
+}
+
+.login-bg {
+  background-image: url('../../assets/img/background.svg');
+  background-size: cover;
+  width: 100%;
+  height: 100vh;
+}
 
 .login-page {
   padding: 48px;
