--- src/pages/LoginSignup/Login/LoginPage.js
@@ -1,21 +1,0 @@
-import React from 'react';
-
-import LoginCard from 'components/LoginCard/LoginCard';
-
-import Logo from 'assets/img/logowhite.svg';
-
-import './LoginPage.scss';
-
-function LoginPage() {
-  return (
-    <div className="login-page">
-      <img src={Logo} alt="snippets logo" />
-      <p className="login-subtitle">
-        An organized, searchable, shareable, and simple code snippets clipboard
-      </p>
-      <LoginCard />
-    </div>
-  );
-}
-
-export default LoginPage;
