--- src/pages/LoginSignup/LoginSignup.scss
@@ -1,13 +1,0 @@
-@import '../../utils/variables.scss';
-
-.login {
-  background-color: $primary-background;
-  height: 100vh;
-}
-
-.login-bg {
-  background-image: url('../../assets/img/background.svg');
-  background-size: cover;
-  width: 100%;
-  height: 100vh;
-}
