--- src/pages/LoginSignup/Signup/index.js
@@ -1,1 +1,0 @@
-export { default } from './SignupPage';
