+++ src/pages/Signup/SignupPage.scss
@@ -1,4 +1,5 @@
-@import '../../../utils/variables.scss';
+@import '../../utils/variables.scss';
+@import '../Login/LoginPage.scss';
 
 .signup-page {
   padding: 48px;
